requirements.txt
selenium
pyyaml
pyvirtualdisplay
